Remove song as user - backend changes
This change are the backend changes to allow users to delete songs THEY
added to the queue. This is done by extending the add function with a
field called "password". This password will be hashed and the hash saved
in the PlaylistEntry data. Further a new call has been added called
DeleteAsUser. Tis requires the password as well. If the passwords match,
a song can be deleted.

This commit also remames the previous "Remove" call to "RemoveAsAdmin" which
works exactly the same way as it has allways done.

It adds a new field to the "Add" function called password.